requirements.txt
Flask==1.1.2
Flask-API==2.0
flask-restx==0.2.0
#Flask-Injector
Flask-SQLAlchemy
bcrypt~=3.2.0
gunicorn
psycopg2
SQLAlchemy~=1.3.20
Flask-JWT-Extended
Werkzeug==1.0.1
#injector~=0.18.4
#APScheduler==3.6.3
PyJWT~=2.1.0
